Python: Split Python docs
This PR will split the Python docs in a separate site. The main reason
for this is that the docs are part of the Java release, which is not in
sync with the Python release cylce. Meaning that there is a high probability
that the docs does not match with current version of the code.

This will publish the docs to Github pages, by pushing this to the `gh-pages`
branch. We can set up an alias from Apache, and point pyiceberg.apache.org to
the github pages endpoint.

I also tried readthedocs, but I found that not straightforward. Mostly because
they have a build process on their end that will pull the code, and build the
docs. This involves another pipeline that we have to monitor, and we have to
set up webhooks. I am a simple man, and I like simple things, therefore I went
for mkdocs. This can push the docs to github pages in a single command:
https://www.mkdocs.org/user-guide/deploying-your-docs/#project-pages

Considerations:

- Decided to keep it to a single page for now, we can break it out into different
  pages later on. Let me know what you think of this.
- We build the docs now when we push to master, probably we'll change this
  later to trigger on tags.
- I've removed the Python docs from the other docs to avoid confusion and make sure
  that we have a single source of truth.

An example is shown here: https://fokko.github.io/incubator-iceberg/
(Once this is merged, I'll remove that one)