Python: Visitor to convert Iceberg to PyArrow schema
This is required for manually specifying datasets:
https://arrow.apache.org/docs/python/dataset.html#manual-specification-of-the-dataset

From PyArrow:

The dataset() function allows easy creation of a Dataset viewing a directory,
crawling all subdirectories for files and partitioning information.
However sometimes discovery is not required and the dataset’s files and
partitions are already known (for example, when this information is stored in metadata).
In this case it is possible to create a Dataset explicitly without any automatic discovery or inference.